Add support for setting window.onload within a script.  Window.onload may be set to the name of a function, e.g., window.onload=initFunc;
To implement this, an onLoad_ member was added to HtmlPage for storing
the new value.  This value is used if set, otherwise the body attribute
value is used as before.  When this value is set within the script, it
is a JavaScript interpreted function object, not a string.  To handle
this function object, the ScriptEngine interface (and test class) as
well as the JavaScriptEngine class were extended to be able to call
function objects.  For debugging these were also extended to convert a
JavaScript object to a string.  The getScope method was factored out to
make it reusable across the new methods.

Window was extended to allow the onload property to be set and read
using the HtmlPage methods.

New test cases were added for reading and setting the onload property
and the change log was updated to describe this fix.

A few missing configuration file items were added to eliminate spurious
messages from the jUnit tests.


git-svn-id: http://svn.code.sf.net/p/htmlunit/code@304 5f5364db-9458-4db8-a492-e30667be6df6